requirements.txt
cognitive_face
django
Pillow